Rewrite as a Slack command
