+++ config.example.js
@@ -1,4 +1,5 @@
 export default {
+  port: 3000,
   githubApiKey: '',
-  slackBotApiKey: '',
+  slackToken: '',
 };
