+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "hacktoberfest-slackbot",
-  "version": "0.1.0",
-  "description": "A Slack bot providing hacktoberfest activity",
+  "name": "hacktoberfest-slack",
+  "version": "0.2.0",
+  "description": "A Slack command providing Hacktoberfest information",
   "main": "index.js",
   "scripts": {
     "start": "babel-node index.js"
@@ -9,8 +9,10 @@
   "author": "Tim Routowicz <troutowicz@gmail.com>",
   "license": "MIT",
   "dependencies": {
-    "botkit": "^0.4.0",
-    "github": "^4.0.0"
+    "body-parser": "^1.15.2",
+    "express": "^4.14.0",
+    "github": "^4.0.0",
+    "morgan": "^1.7.0"
   },
   "devDependencies": {
     "babel-cli": "^6.16.0",
